fix: hide share button in git mode
