+++ e2e/tests/share.spec.ts
@@ -6,18 +6,17 @@ async function loadPage(page: Page) {
 }
 
 // ============================================================
-// Share Feature — Git Mode (default share URL: crit.live)
+// Share Feature — Git Mode (share button hidden)
 // ============================================================
-test.describe('Share — Default Configuration', () => {
-  test('share button is visible by default', async ({ page }) => {
+test.describe('Share — Git Mode', () => {
+  test('share button is hidden in git mode', async ({ page }) => {
     await loadPage(page);
 
     const shareBtn = page.locator('#shareBtn');
-    await expect(shareBtn).toBeVisible();
-    await expect(shareBtn).toHaveText('Share');
+    await expect(shareBtn).toBeHidden();
   });
 
-  test('config API returns default share_url', async ({ request }) => {
+  test('config API still returns share_url', async ({ request }) => {
     const res = await request.get('/api/config');
     const config = await res.json();
     expect(config.share_url).toBeTruthy();
@@ -33,19 +32,5 @@ test.describe('Share — Default Configuration', () => {
     const res = await request.get('/api/config');
     const config = await res.json();
     expect(config.delete_token).toBe('');
-  });
-
-  test('no share-related toasts are shown on initial load', async ({ page }) => {
-    await loadPage(page);
-
-    const shareToast = page.locator('#toast-share');
-    await expect(shareToast).toHaveCount(0);
-  });
-
-  test('share button is enabled and clickable', async ({ page }) => {
-    await loadPage(page);
-
-    const shareBtn = page.locator('#shareBtn');
-    await expect(shareBtn).toBeEnabled();
   });
 });
